Updated Manual
We do not need an observable interface for broadcast as there is a built-in OnReceive method meant for BLE broadcasting.  Kindly look at the 1st page of the manual for more information.  If there's any problem, please signal it by creating an issue in the git or message me directly.